refactor(admin-cli): remove dead rpc CLI output helpers
Remove unused common CLI output helpers from the rpc crate, including
dead database helpers and generic formatted-output/table traits. Move
the remaining destination/output formatting code into admin-cli and keep
VPC prefix table rendering local to its only remaining use site.
