+++ crates/admin-cli/src/vpc_prefix/create/cmd.rs
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-use ::rpc::admin_cli::output::{FormattedOutput, OutputFormat};
+use ::rpc::admin_cli::output::OutputFormat;
 
 use super::args::Args;
 use crate::errors::{CarbideCliError, CarbideCliResult};
@@ -34,6 +34,6 @@ pub async fn create(
         .map(ShowOutput::One)?;
 
     output
-        .write_output(output_format, ::rpc::admin_cli::Destination::Stdout())
+        .write_output(output_format, crate::Destination::Stdout())
         .map_err(CarbideCliError::from)
 }
